requirements.txt
fire
ftfy
nltk
tqdm
pytz
munch
seaborn
notebook
matplotlib
scikit-learn
tensorboard
wandb
openai
google-cloud-aiplatform
langchain
transformers==4.45.2
diffusers==0.30.3
invisible_watermark
safetensors
peft
timm
loralib
open_clip_torch
compel
decord
bitsandbytes
git+https://github.com/openai/CLIP.git
